% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ages2stages}
\alias{ages2stages}
\title{A data frame mapping Southern Resident killer whales ages and sexes to life stages}
\format{
A data frame.
}
\usage{
ages2stages
}
\description{
A data frame mapping Southern Resident killer whales ages and sexes to life stages
}
\keyword{datasets}
