% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orca}
\alias{orca}
\title{Example census for Southern Resident killer whales}
\format{
A data frame.
}
\usage{
orca
}
\description{
Example census for Southern Resident killer whales
}
\keyword{datasets}
