% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Function to project SRKW population in the future using individual based models}
\usage{
project(
  whale_data,
  seed = 123,
  n_years = 30,
  scenarios = c("1981:2020"),
  n_iter = 200,
  verbose = FALSE,
  p_female = NA,
  n_births = NA
)
}
\arguments{
\item{whale_data}{the expanded data frame, returned by \code{expand()}}

\item{seed}{the random seed used to initialize simulations, defaults to \code{123}}

\item{n_years}{the length of the time series projections, defaults to 30 time steps}

\item{scenarios}{a character string, or vector of strings describing which scenario
to be run. Should be entered as a range of years separated by a colon,
e.g. "1981:1985" to use rates from 1981 to 1985 or ("1981:1985","1981:2021")}

\item{n_iter}{the number of iterations per scenario, defaults to 200}

\item{verbose}{whether to print status updates to screen, defaults to FALSE}

\item{p_female}{The probability of a female birth, defaults to the empirical value since 1980}

\item{n_births}{The number of births used to calculate \code{p_female}. By default, this defaults to the empirical
value observed since 1980 the probability of female birth. Larger values result in more precise estimates of \code{p_female}}
}
\description{
Function to project SRKW population in the future using individual based models
}
