% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projections}
\alias{projections}
\title{A data frame containing future SRKW population projections}
\format{
A data frame.
}
\usage{
projections
}
\description{
A data frame containing future SRKW population projections
}
\keyword{datasets}
