% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedata.R
\name{writedata}
\alias{writedata}
\title{Function takes csv file and converts is to saved object}
\usage{
writedata(filename)
}
\arguments{
\item{filename}{A file location (csv file) with births and deaths of each animal}
}
\description{
Function takes csv file and converts is to saved object
}
\examples{
\dontrun{
writedata("inst/extdata/whaleData_08-06-2018.csv")
}
}
