% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Function expand expands a data frame from animals as observations to animal-years as obserations}
\usage{
expand(
  filename,
  start_year = 1976,
  current_year = NULL,
  fem_age_mat = 10,
  fem_age_senesc = 43
)
}
\arguments{
\item{filename}{A file location (csv file) with births and deaths of each animal}

\item{start_year}{Starting year for the expanded data, starts at 1976}

\item{current_year}{End year for the expanded data, defaults to current calendar year}

\item{fem_age_mat}{Female age at maturity, 10 years of age by default}

\item{fem_age_senesc}{Female age at reproductive senescence, defaults to 43}
}
\description{
Function expand expands a data frame from animals as observations to animal-years as obserations
}
\examples{
\dontrun{
library(kwdemog)
data(orca)
expanded_data = expand(orca)
# or use a filename
expanded_data = expand("use_this_file.csv", start_year = 1979, current_year = 2018,
fem_age_mat = 10, fem_age_senesc = 43)
}
}
